ai-banking-survey-analyzer/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
boto3==1.42.73
botocore==1.42.73
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.128.8
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.28
langchain-aws==0.2.35
langchain-core==0.3.83
langchain-text-splitters==0.3.11
langsmith==0.4.37
numpy==1.26.4
orjson==3.11.5
packaging==25.0
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.49.3
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
uuid_utils==0.14.1
uvicorn==0.39.0
zstandard==0.25.0
